pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-atc-optimization"
version = "1.0.0"
description = "OpenEnv benchmark for realistic air traffic control disruption recovery"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "openai>=2.30.0",
    "openenv-core[core]>=0.2.3",
    "pydantic>=2.12.0",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "requests>=2.33.0",
]
training = [
    "trl>=0.9.6",
    "datasets>=2.20.0",
    "transformers>=4.43.0",
    "torch>=2.3.0",
    "accelerate>=0.32.0",
    "bitsandbytes>=0.43.0",
    "peft>=0.12.0",
    "matplotlib>=3.9.0",
    "numpy>=1.26.0",
]
training-unsloth = [
    "unsloth[colab-new]>=2024.8",
    "trl>=0.9.6",
    "datasets>=2.20.0",
    "matplotlib>=3.9.0",
    "numpy>=1.26.0",
]

[project.scripts]
server = "atc_env.server.app:main"
atc-train = "training.train_grpo:main"
atc-eval = "training.eval:main"
atc-plot = "training.plot_rewards:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["atc_env*", "multi_agent*", "training*", "server*"]
